refactor: 패키지명 소문자로 변경(Lotto -> lotto, StringCalculator -> stringCalculator)
