--- src/main/java/Lotto/domain/RankMatcher.java
@@ -1,5 +1,0 @@
-package Lotto.domain;
-
-public interface RankMatcher {
-    boolean  match(int countOfMatch, boolean matchBonus);
-}
