--- src/main/java/Lotto/view/input/InputViewInterface.java
@@ -1,16 +1,0 @@
-package Lotto.view.input;
-
-import Lotto.domain.LottoNumber;
-
-import java.util.List;
-import java.util.function.Predicate;
-
-public interface InputViewInterface {
-    int getNumberInput();
-
-    int[] getNumberListInput();
-
-    int getValidatedNumberInput(Predicate<Integer> validator, String errorMessage);
-
-    List<List<LottoNumber>> getManualLottoNumbers(int count);
-}
