--- src/main/java/StringCalculator/domain/Operand.java
@@ -1,33 +1,0 @@
-package StringCalculator.domain;
-
-import java.util.Objects;
-
-public class Operand {
-    private final int value;
-
-    public Operand(int value) {
-        this.value = value;
-    }
-
-    public int getValue() {
-        return value;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Operand operand = (Operand) o;
-        return value == operand.value;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(value);
-    }
-
-    @Override
-    public String toString() {
-        return String.valueOf(value);
-    }
-}
