--- src/main/java/StringCalculator/view/ResultView.java
@@ -1,7 +1,0 @@
-package StringCalculator.view;
-
-public class ResultView {
-    public static void printResult(int result) {
-        System.out.println(result);
-    }
-}
