+++ src/main/java/lotto/domain/RankMatcher.java
@@ -1,0 +1,5 @@
+package lotto.domain;
+
+public interface RankMatcher {
+    boolean  match(int countOfMatch, boolean matchBonus);
+}
