+++ src/main/java/lotto/view/input/InputViewInterface.java
@@ -1,0 +1,16 @@
+package lotto.view.input;
+
+import lotto.domain.LottoNumber;
+
+import java.util.List;
+import java.util.function.Predicate;
+
+public interface InputViewInterface {
+    int getNumberInput();
+
+    int[] getNumberListInput();
+
+    int getValidatedNumberInput(Predicate<Integer> validator, String errorMessage);
+
+    List<List<LottoNumber>> getManualLottoNumbers(int count);
+}
