+++ src/main/java/lotto/view/output/OutputViewInterface.java
@@ -1,0 +1,15 @@
+package lotto.view.output;
+
+import lotto.domain.Lottos;
+import lotto.domain.ResultStats;
+
+public interface OutputViewInterface {
+
+    void printPrompt(String prompt);
+
+    void printTicketCount(int manualCount, int autoCount);
+
+    void printLottos(Lottos lottos);
+
+    void printResultStats(ResultStats resultStats, int totalSpent);
+}
