+++ src/main/java/stringCalculator/view/ResultView.java
@@ -1,0 +1,7 @@
+package stringCalculator.view;
+
+public class ResultView {
+    public static void printResult(int result) {
+        System.out.println(result);
+    }
+}
