+++ src/test/java/LottoNumberTest.java
@@ -1,4 +1,4 @@
-import Lotto.domain.LottoNumber;
+import lotto.domain.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
